@@ -16,50 +16,51 @@
         ]
     },
     "require": {
-        "php": ">=7.3 <7.5",
+        "php": ">=7.4.0 <8.1",
         "ext-zlib": "^7.2",
         "ext-zip": "^1.15",
-        "symfony/console": "~3.4.0",
-        "symfony/debug": "~3.4.0",
-        "symfony/form": "~3.4.0",
-        "symfony/intl": "~3.4.0",
-        "symfony/options-resolver": "~3.4.0",
-        "symfony/process": "~3.4.0",
-        "symfony/security": "~3.4.0",
-        "symfony/translation": "~3.4.0",
-        "symfony/validator": "~3.4.0",
-        "symfony/yaml": "~3.4.0",
-        "symfony/property-access": "~3.4.0",
-        "symfony/cache": "~3.4.0",
-        "symfony/asset": "~3.4.0",
-        "symfony/class-loader": "~3.4.0",
-        "symfony/dependency-injection": "~3.4.0",
-        "symfony/config": "~3.4.0",
-        "symfony/event-dispatcher": "~3.4.0",
-        "symfony/finder": "~3.4.0",
-        "symfony/http-foundation": "~3.4.0",
-        "symfony/http-kernel": "~3.4.0",
-        "symfony/filesystem": "~3.4.0",
-        "symfony/routing": "~3.4.0",
-        "symfony/security-core": "~3.4.0",
-        "symfony/security-csrf": "~3.4.0",
-        "symfony/stopwatch": "~3.4.0",
-        "symfony/templating": "~3.4.0",
-        "symfony/dotenv": "~3.4.0",
-        "symfony/expression-language": "~3.4.0",
-        "symfony/framework-bundle": "~3.4.0",
+        "ext-imap": "*",
+        "symfony/console": "~4.4.0",
+        "symfony/debug": "~4.4.0",
+        "symfony/form": "~4.4.0",
+        "symfony/intl": "~4.4.0",
+        "symfony/options-resolver": "~4.4.0",
+        "symfony/process": "~4.4.0",
+        "symfony/security": "~4.4.0",
+        "symfony/translation": "~4.4.0",
+        "symfony/validator": "~4.4.0",
+        "symfony/yaml": "~4.4.0",
+        "symfony/property-access": "~4.4.0",
+        "symfony/cache": "~4.4.0",
+        "symfony/asset": "~4.4.0",
+        "symfony/dependency-injection": "~4.4.0",
+        "symfony/config": "~4.4.0",
+        "symfony/event-dispatcher": "~4.4.0",
+        "symfony/finder": "~4.4.0",
+        "symfony/http-foundation": "~4.4.0",
+        "symfony/http-kernel": "~4.4.0",
+        "symfony/filesystem": "~4.4.0",
+        "symfony/routing": "~4.4.0",
+        "symfony/security-core": "~4.4.0",
+        "symfony/security-csrf": "~4.4.0",
+        "symfony/stopwatch": "~4.4.0",
+        "symfony/templating": "~4.4.0",
+        "symfony/dotenv": "~4.4.0",
+        "symfony/expression-language": "~4.4.0",
+        "symfony/framework-bundle": "~4.4.0",
         "symfony/monolog-bundle": "~3.5.0",
-        "symfony/security-bundle": "~3.4.0",
-        "symfony/twig-bundle": "~3.4.0",
-        "symfony/doctrine-bridge": "~3.4.0",
-        "symfony/monolog-bridge": "~3.4.0",
+        "symfony/security-bundle": "~4.4.0",
+        "symfony/twig-bundle": "~4.4.0",
+        "symfony/doctrine-bridge": "~4.4.0",
+        "symfony/monolog-bridge": "~4.4.0",
         "symfony/swiftmailer-bundle": "~3.3",
         "symfony/polyfill-mbstring": "~1.0",
         "symfony/security-acl": "~3.0.0",
         "symfony/polyfill-php73": "^1.13",
         "symfony/polyfill-php74": "^1.13",
+        "symfony/css-selector": "~4.4.0",
+        "symfony/http-client": "~4.4.0",
 
-        "sensio/distribution-bundle": "~5.0",
         "sensio/framework-extra-bundle": "~3.0",
 
         "doctrine/doctrine-bundle": "~1.12.0",
@@ -111,16 +108,21 @@
         "bandwidth-throttle/token-bucket": "^2.0",
         "gaufrette/extras": "^0.1.0",
         "gaufrette/aws-s3-adapter": "^0.4.0",
+        "theofidry/psysh-bundle": "~4.3.0",
+        "matomo/device-detector": "^4.0",
+        "symfony/lock": "^5.1",
+        "composer/package-versions-deprecated": "1.11.99.1",
         "phpoffice/phpspreadsheet": "^1.15",
         "predis/predis": "^1.1",
         "composer/installers": "^1.10",
         "mautic/grapejs-builder": "1.0-RC4"
     },
     "require-dev": {
-        "symfony/web-profiler-bundle": "~3.4.0",
-        "symfony/var-dumper": "~3.4.0",
-        "symfony/browser-kit": "~3.4.0",
-        "symfony/dom-crawler": "~3.4.0",
+        "symfony/web-profiler-bundle": "~4.4.0",
+        "symfony/var-dumper": "~4.4.0",
+        "symfony/browser-kit": "~4.4.0",
+        "symfony/dom-crawler": "~4.4.0",
+
         "mautic/transifex": "^4.0",
         "http-interop/http-factory-guzzle": "^1.0",
         "webfactory/exceptions-bundle": "~4.3",
