@@ -13,8 +13,8 @@ namespace Mautic\ApiBundle\DependencyInjection\Factory;
 
 use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
 use Symfony\Component\Config\Definition\Builder\NodeDefinition;
+use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\DefinitionDecorator;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
@@ -29,12 +29,12 @@ class ApiFactory implements SecurityFactoryInterface
     {
         $providerId = 'security.authentication.provider.mautic_api.'.$id;
         $container
-            ->setDefinition($providerId, new DefinitionDecorator('mautic_api.security.authentication.provider'))
+            ->setDefinition($providerId, new ChildDefinition('mautic_api.security.authentication.provider'))
             ->replaceArgument(0, new Reference($userProvider))
         ;
 
         $listenerId = 'security.authentication.listener.mautic_api.'.$id;
-        $container->setDefinition($listenerId, new DefinitionDecorator('mautic_api.security.authentication.listener'));
+        $container->setDefinition($listenerId, new ChildDefinition('mautic_api.security.authentication.listener'));
 
         return [$providerId, $listenerId, $defaultEntryPoint];
     }
