@@ -30,13 +30,9 @@ abstract class AbstractStandardFormController extends AbstractFormController
     /**
      * Support non-index pages such as modal forms.
      *
-     * @param string $route
-     * @param array  $parameters
-     * @param int    $referenceType
-     *
      * @return bool|string
      */
-    public function generateUrl($route, $parameters = [], $referenceType = UrlGeneratorInterface::ABSOLUTE_PATH)
+    protected function generateUrl(string $route, array $parameters = [], int $referenceType = UrlGeneratorInterface::ABSOLUTE_PATH): string
     {
         if (false === $route) {
             return false;
