@@ -11,9 +11,12 @@
 
 namespace Mautic\CoreBundle\DependencyInjection;
 
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\Alias;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
+use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\ExpressionLanguage\Expression;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
@@ -31,6 +34,11 @@ class MauticCoreExtension extends Extension
      */
     public function load(array $configs, ContainerBuilder $container)
     {
+        // Auto-wire commands to keep support for the M3 way although best practice is to register each command
+        // as a service and tag with console.command or include in a Mautic config.php services[command] array.
+        $loader = new PhpFileLoader($container, new FileLocator(__DIR__.'/../../../config'));
+        $loader->load('services.php');
+
         $bundles = array_merge($container->getParameter('mautic.bundles'), $container->getParameter('mautic.plugin.bundles'));
 
         // Store menu renderer options to create unique renderering classes per menu
@@ -67,7 +75,11 @@ class MauticCoreExtension extends Extension
                             $defaultTag = 'mautic.integration';
                             break;
                         case 'command':
+                        case 'commands':
                             $defaultTag = 'console.command';
+                            break;
+                        case 'controllers':
+                            $defaultTag = 'controller.service_arguments';
                             break;
                         default:
                             $defaultTag = false;
@@ -100,19 +112,21 @@ class MauticCoreExtension extends Extension
                         }
 
                         // Set service alias
+                        $alias = new Alias($name);
+                        $alias->setPublic(true);
                         if (isset($details['serviceAlias'])) {
                             // Fix escaped sprintf placeholders
                             $details['serviceAlias'] = str_replace('%%', '%', $details['serviceAlias']);
-                            $container->setAlias(sprintf($details['serviceAlias'], $name), $name);
+                            $container->setAlias(sprintf($details['serviceAlias'], $name), $alias);
                         } elseif (isset($details['serviceAliases'])) {
-                            foreach ($details['serviceAliases'] as $alias) {
-                                $alias = str_replace('%%', '%', $alias);
-                                $container->setAlias(sprintf($alias, $name), $name);
+                            foreach ($details['serviceAliases'] as $aliasName) {
+                                $aliasName = str_replace('%%', '%', $aliasName);
+                                $container->setAlias(sprintf($aliasName, $name), $alias);
                             }
                         }
-                        // Alias with class name
+                        // Symfony 4 is requiring the classname for some auto-wired services (controllers)
                         if ($name !== $details['class']) {
-                            $container->setAlias($details['class'], $name);
+                            $container->setAlias($details['class'], $alias);
                         }
 
                         // Generate definition arguments
@@ -133,10 +147,6 @@ class MauticCoreExtension extends Extension
                             $definitionArguments
                         ));
 
-                        if (isset($details['public'])) {
-                            $definition->setPublic($details['public']);
-                        }
-
                         // Generate tag and tag arguments
                         if (isset($details['tags'])) {
                             $tagArguments = (!empty($details['tagArguments'])) ? $details['tagArguments'] : [];
@@ -152,7 +162,7 @@ class MauticCoreExtension extends Extension
                                 $definition->addTag($tag, $tagArguments[$k]);
 
                                 if ('mautic.email_transport' === $tag) {
-                                    $container->setAlias(sprintf('swiftmailer.mailer.transport.%s', $name), $name);
+                                    $container->setAlias(sprintf('swiftmailer.mailer.transport.%s', $name), $alias);
                                 }
                             }
                         } else {
@@ -167,7 +177,7 @@ class MauticCoreExtension extends Extension
                                 $definition->addTag($tag, $tagArguments);
 
                                 if ('mautic.email_transport' === $tag) {
-                                    $container->setAlias(sprintf('swiftmailer.mailer.transport.%s', $name), $name);
+                                    $container->setAlias(sprintf('swiftmailer.mailer.transport.%s', $name), $alias);
                                 }
                             }
 
@@ -176,10 +186,9 @@ class MauticCoreExtension extends Extension
                             }
                         }
 
-                        // Set public service
-                        if (!empty($details['public'])) {
-                            $definition->setPublic($details['public']);
-                        }
+                        // Default to a public service
+                        $public = $details['public'] ?? true;
+                        $definition->setPublic($public);
 
                         // Set lazy service
                         if (!empty($details['lazy'])) {
@@ -204,28 +213,6 @@ class MauticCoreExtension extends Extension
                         // Set service configurator
                         if (!empty($details['configurator'])) {
                             $definition->setConfigurator($details['configurator']);
-                        }
-
-                        // Set scope - Deprecated as of Symfony 2.8 and removed in 3.0
-                        if (!empty($details['scope'])) {
-                            $definition->setScope($details['scope']);
-                        } elseif ('templating' == $type) {
-                            $definition->setScope('request');
-                        }
-
-                        // Set factory service - Deprecated as of Symfony 2.6 and removed in Symfony 3.0
-                        if (!empty($details['factoryService'])) {
-                            $definition->setFactoryService($details['factoryService']);
-                        }
-
-                        // Set factory class - Deprecated as of Symfony 2.6 and removed in Symfony 3.0
-                        if (!empty($details['factoryClass'])) {
-                            $definition->setFactoryClass($details['factoryClass']);
-                        }
-
-                        // Set factory method - Deprecated as of Symfony 2.6 and removed in Symfony 3.0
-                        if (!empty($details['factoryMethod'])) {
-                            $definition->setFactoryMethod($details['factoryMethod']);
                         }
 
                         // Set factory - Preferred API since Symfony 2.6
