@@ -14,7 +14,6 @@ namespace Mautic\CoreBundle\Doctrine\Type;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\ConversionException;
 use Mautic\CoreBundle\Helper\UTF8Helper;
-use Symfony\Component\Debug\Exception\ContextErrorException;
 
 /**
  * Type that maps a PHP array to a clob SQL type.
@@ -82,7 +81,7 @@ class ArrayType extends \Doctrine\DBAL\Types\ArrayType
             return $value;
         } catch (ConversionException $exception) {
             return [];
-        } catch (ContextErrorException $exeption) {
+        } catch (\ErrorException $exeption) {
             return [];
         }
     }
