@@ -633,8 +633,12 @@ return [
             'mautic.guzzle.client.factory' => [
                 'class' => \Mautic\EmailBundle\Swiftmailer\Guzzle\ClientFactory::class,
             ],
+            /**
+             * Needed for Sparkpost integration. Can be removed when this integration is moved to
+             * its own plugin.
+             */
             'mautic.guzzle.client' => [
-                'class'     => \Http\Adapter\Guzzle6\Client::class,
+                'class'     => \Http\Adapter\Guzzle7\Client::class,
                 'factory'   => ['@mautic.guzzle.client.factory', 'create'],
             ],
             'mautic.helper.mailbox' => [
