@@ -17,7 +17,7 @@ class PageControllerTest extends MauticMysqlTestCase
     protected function setUp(): void
     {
         parent::setUp();
-        $this->prefix = $this->container->getParameter('mautic.db_table_prefix');
+        $this->prefix = self::$container->getParameter('mautic.db_table_prefix');
     }
 
     public function testLandingPageTracking()
