@@ -3,7 +3,7 @@
 namespace Mautic\WebhookBundle\Tests\Functional;
 
 use GuzzleHttp\Psr7\Response as GuzzleResponse;
-use Http\Adapter\Guzzle6\Client;
+use Http\Mock\Client;
 use Mautic\CoreBundle\Test\MauticMysqlTestCase;
 use Mautic\WebhookBundle\Command\ProcessWebhookQueuesCommand;
 use Mautic\WebhookBundle\Entity\Event;
@@ -13,6 +13,7 @@ use Mautic\WebhookBundle\Entity\WebhookQueueRepository;
 use Mautic\WebhookBundle\Model\WebhookModel;
 use PHPUnit\Framework\Assert;
 use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -44,7 +45,7 @@ class WebhookFunctionalTest extends MauticMysqlTestCase
         $httpClient                    = new class() extends Client {
             public $sendRequestCounter = 0;
 
-            public function sendRequest(RequestInterface $request)
+            public function sendRequest(RequestInterface $request): ResponseInterface
             {
                 Assert::assertSame('://whatever.url', $request->getUri()->getPath());
                 $jsonPayload = json_decode($request->getBody()->getContents(), true);
@@ -56,7 +57,7 @@ class WebhookFunctionalTest extends MauticMysqlTestCase
             }
         };
 
-        $this->container->set('mautic.guzzle.client', $httpClient);
+        self::$container->set('mautic.guzzle.client', $httpClient);
 
         /** @var WebhookQueueRepository $webhookQueueRepository */
         $webhookQueueRepository = $this->em->getRepository(WebhookQueue::class);
