@@ -13,7 +13,7 @@ namespace Mautic\WebhookBundle\Tests\Http;
 
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
-use Http\Adapter\Guzzle6\Client as GuzzleClient;
+use Http\Mock\Client as GuzzleClient;
 use Mautic\CoreBundle\Helper\CoreParametersHelper;
 use Mautic\WebhookBundle\Http\Client;
 use PHPUnit\Framework\TestCase;
