@@ -5,6 +5,8 @@ use MauticPlugin\MauticCrmBundle\Tests\Pipedrive\Mock\Client;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\Dotenv\Dotenv;
 
+/** @var \Symfony\Component\DependencyInjection\ContainerBuilder $container */
+
 /*
  * @copyright   2014 Mautic Contributors. All rights reserved
  * @author      Mautic
@@ -44,7 +46,9 @@ $container->loadFromExtension('framework', [
 
 $container->setParameter('mautic.famework.csrf_protection', true);
 
-$container->register('mautic_integration.pipedrive.guzzle.client', Client::class);
+$container
+    ->register('mautic_integration.pipedrive.guzzle.client', Client::class)
+    ->setPublic(true);
 
 $container->loadFromExtension('web_profiler', [
     'toolbar'             => false,
@@ -140,17 +144,21 @@ $container->setParameter('mautic.batch_sleep_time', 0);
 // Turn off creating of indexes in lead field fixtures
 $container->register('mautic.install.fixture.lead_field', \Mautic\InstallBundle\InstallFixtures\ORM\LeadFieldData::class)
     ->addArgument(false)
-    ->addTag(FixturesCompilerPass::FIXTURE_TAG);
+    ->addTag(FixturesCompilerPass::FIXTURE_TAG)
+    ->setPublic(true);
 $container->register('mautic.lead.fixture.contact_field', \Mautic\LeadBundle\DataFixtures\ORM\LoadLeadFieldData::class)
     ->addArgument(false)
-    ->addTag(FixturesCompilerPass::FIXTURE_TAG);
+    ->addTag(FixturesCompilerPass::FIXTURE_TAG)
+    ->setPublic(true);
 
 // Use static namespace for token manager
 $container->register('security.csrf.token_manager', \Symfony\Component\Security\Csrf\CsrfTokenManager::class)
     ->addArgument(new Reference('security.csrf.token_generator'))
     ->addArgument(new Reference('security.csrf.token_storage'))
-    ->addArgument('test');
+    ->addArgument('test')
+    ->setPublic(true);
 
 // Stub HTTP client to prevent accidental request to third parties
 $container->register('mautic.http.client', \GuzzleHttp\Client::class)
-    ->setFactory('\Mautic\CoreBundle\Test\Guzzle\ClientFactory::stub');
+    ->setFactory('\Mautic\CoreBundle\Test\Guzzle\ClientFactory::stub')
+    ->setPublic(true);
