@@ -107,7 +107,7 @@ class PushDataToPipedriveCommandTest extends PipedriveTest
 
     private function executeCommand()
     {
-        $kernel      = $this->container->get('kernel');
+        $kernel      = self::$container->get('kernel');
         $application = new Application($kernel);
         $application->setAutoExit(false);
 
